Updated Gig model
Moved Gig enums to new file
Separated GigCategories into categories and subcategories
Gig now hold data about client and provider accounts for transactions processing
